auto trust level changes with user type
Instead of one auto trust level setting for all, we have separate settings for the 4 user
types now. We usually want to have lower trust levels for guest users than
for eligible members, for example.
If the user type changes after the user has registered, the level is
set, even to lower levels. How does Discourse react in this case? We
need to test this.